==> script/01_database_building/03_license_based_indices/02_innovations/04_classify_crop_qualities.R <==
# run the keras model on... 
# pp_db (need to save from 02)
# pvp_db (need to save from 02)

# save as an output to read in for who breeds what (replace pvp_db and pp_db)
# Load in the keras package
library(keras)

# Install TensorFlow
tensorflow::install_tensorflow()
imdb <- dataset_imdb()
x_train <- imdb$train$x
y_train <- imdb$train$y
x_test <- imdb$test$x
y_test <- imdb$test$y

codes <- readRDS('data/codes.RDS')
